requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
playwright==1.44.0
trafilatura==1.9.0
beautifulsoup4==4.12.3
lxml==5.1.0
numpy==1.26.4
sqlalchemy==2.0.30
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
